deps/acoustic-engine/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "acoustic-engine"
version = "1.0.0"
description = "Real-time acoustic alarm pattern detection engine"
readme = "README.md"
license = {text = "CC BY-NC 4.0"}
authors = [
    {name = "h0tp", email = "try-cycling-reverb@duck.com"}
]
keywords = ["audio", "alarm", "detection", "fft", "pattern-matching", "iot", "home-automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Analysis",
    "Topic :: Home Automation",
]
requires-python = ">=3.9"
dependencies = [
    "numpy>=1.21.0",
    "scipy>=1.7.0",
    "PyAudio>=0.2.11",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1.0",
]
audio = [
    "pydub>=0.25.0",  # For MP3, OGG, FLAC support (requires ffmpeg)
]
tuner = [
    # Tuner uses built-in http.server, no extra deps needed
]

[project.urls]
Homepage = "https://github.com/h0tp-ftw/acoustic-engine"
Documentation = "https://github.com/h0tp-ftw/acoustic-engine#readme"
Repository = "https://github.com/h0tp-ftw/acoustic-engine"
Issues = "https://github.com/h0tp-ftw/acoustic-engine/issues"

[project.scripts]
alarm-tuner = "acoustic_engine.tuner.__main__:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"acoustic_engine.tuner" = ["*.html", "*.css", "*.js"]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
